+++ src/Sound/Tidal/Tempo.hs
@@ -96,20 +96,26 @@ clocked config callback
   where loop tempoMV st =
           do -- putStrLn $ show $ nowArc ts
              tempo <- readMVar tempoMV               
+             t <- O.time
              let frameTimespan = cFrameTimespan config
-                 -- 'now' comes from clock ticks, nothing to do with cycles
                  logicalT ticks' = start st + fromIntegral ticks' * frameTimespan
                  logicalNow = logicalT $ ticks st + 1
-                 -- the tempo is just used to convert logical time to cycles
+                 -- Wait maximum of two frames
+                 delta = min (frameTimespan * 2) (logicalNow - t)
                  e = timeToCycles tempo logicalNow
                  s = if starting st && synched tempo
                      then timeToCycles tempo (logicalT $ ticks st)
                      else P.stop $ nowArc st
-                 st' = st {ticks = ticks st + 1, nowArc = P.Arc s e,
+             when (t < logicalNow) $ threadDelay (floor $ delta * 1000000)
+             t' <- O.time
+             let actualTick = floor $ (t' - start st) / frameTimespan
+                 -- reset ticks if ahead/behind by 4 or more
+                 newTick | (abs $ actualTick - ticks st) > 4 = actualTick
+                         | otherwise = (ticks st) + 1
+                 st' = st {ticks = newTick,
+                           nowArc = P.Arc s e,
                            starting = not (synched tempo)
                           }
-             t <- O.time
-             when (t < logicalNow) $ threadDelay (floor $ (logicalNow - t) * 1000000)
              callback tempoMV st'
              loop tempoMV st'
 
@@ -120,7 +126,7 @@ clientListen config s =
          hostname = cTempoAddr config
          port = cTempoPort config
      (remote_addr:_) <- N.getAddrInfo Nothing (Just hostname) Nothing
-     local <- O.udpServer "127.0.0.1" tempoClientPort
+     local <- O.udpServer "0.0.0.0" tempoClientPort
      let (N.SockAddrInet _ a) = N.addrAddress remote_addr
          remote = N.SockAddrInet (fromIntegral port) a
          t = defaultTempo s local remote