+++ tidal.el
@@ -57,6 +57,14 @@
   ()
   "*Arguments to the haskell interpreter (default=none).")
 
+(defvar tidal-boot-script-path
+  (concat (substring
+           (shell-command-to-string
+            "ghc-pkg describe $(ghc-pkg latest tidal) | grep data-dir | cut -f2 -d' '") 0 -1)
+          "/BootTidal.hs")
+  "*Full path to BootTidal.hs (inferred by introspecting ghc-pkg package db)."
+  )
+
 (defvar tidal-literate-p
   t
   "*Flag to indicate if we are in literate mode (default=t).")
@@ -87,60 +95,7 @@
      nil
      tidal-interpreter-arguments)
     (tidal-see-output))
-  (tidal-send-string ":set prompt \"\"")
-  (if (string< tidal-interpreter-version "8.2.0")
-      (tidal-send-string ":set prompt2 \"\"")
-    (tidal-send-string ":set prompt-cont \"\""))
-  (tidal-send-string ":set -XOverloadedStrings
-import Sound.Tidal.Context
-tidal <- startMulti [superdirtTarget {oLatency = 0.1, oAddress = \"127.0.0.1\", oPort = 57120}] (defaultConfig {cFrameTimespan = 1/20, cCtrlListen = True})
-let p = streamReplace tidal
-    hush = streamHush tidal
-    list = streamList tidal
-    mute = streamMute tidal
-    unmute = streamUnmute tidal
-    solo = streamSolo tidal
-    unsolo = streamUnsolo tidal
-    once = streamOnce tidal False
-    asap = streamOnce tidal True
-    nudgeAll = streamNudgeAll tidal
-    all = streamAll tidal
-    resetCycles = streamResetCycles tidal
-    setcps = asap . cps
-    xfade i = transition tidal (Sound.Tidal.Transition.xfadeIn 4) i
-    xfadeIn i t = transition tidal (Sound.Tidal.Transition.xfadeIn t) i
-    histpan i t = transition tidal (Sound.Tidal.Transition.histpan t) i
-    wait i t = transition tidal (Sound.Tidal.Transition.wait t) i
-    waitT i f t = transition tidal (Sound.Tidal.Transition.waitT f t) i
-    jump i = transition tidal (Sound.Tidal.Transition.jump) i
-    jumpIn i t = transition tidal (Sound.Tidal.Transition.jumpIn t) i
-    jumpIn' i t = transition tidal (Sound.Tidal.Transition.jumpIn' t) i
-    jumpMod i t = transition tidal (Sound.Tidal.Transition.jumpMod t) i
-    mortal i lifespan release = transition tidal (Sound.Tidal.Transition.mortal lifespan release) i
-    interpolate i = transition tidal (Sound.Tidal.Transition.interpolate) i
-    interpolateIn i t = transition tidal (Sound.Tidal.Transition.interpolateIn t) i
-    clutch i = transition tidal (Sound.Tidal.Transition.clutch) i
-    clutchIn i t = transition tidal (Sound.Tidal.Transition.clutchIn t) i
-    anticipate i = transition tidal (Sound.Tidal.Transition.anticipate) i
-    anticipateIn i t = transition tidal (Sound.Tidal.Transition.anticipateIn t) i
-    d1 = p 1
-    d2 = p 2 -- . (|< orbit 1)
-    d3 = p 3 -- . (|< orbit 2)
-    d4 = p 4 -- . (|< orbit 3)
-    d5 = p 5 -- . (|< orbit 4)
-    d6 = p 6 -- . (|< orbit 5)
-    d7 = p 7 -- . (|< orbit 6)
-    d8 = p 8 -- . (|< orbit 7)
-    d9 = p 9 -- . (|< orbit 8)
-    d10 = p 10
-    d11 = p 11
-    d12 = p 12
-    d13 = p 13
-    d14 = p 14
-    d15 = p 15
-    d16 = p 16
-  ")
-  (tidal-send-string ":set prompt \"tidal> \"")
+  (tidal-send-string (concat ":script " tidal-boot-script-path))
 )
 
 (defun tidal-see-output ()