+++ CHANGELOG.md
@@ -1,5 +1,55 @@
 # TidalCycles log of changes
 
+## 1.0.11 - as yet unnamed
+
+2019-04-17  Alex McLean  <alex@slab.org>
+	* Add `bite` function for slicing patterns (rather than samples)
+	* Tweak tidal.el to attempt to infer location of default BootTidal.hs
+	* Skip time (forward or backward) if the reference clock jumps suddenly
+	* Fix `fit` - @bgold-cosmos
+	* Remove 'asap'
+	* Add cB for boolean control input
+	* `pickF` for choosing between functions with a pattern of integers
+	* `select` for choosing between list of patterns with a floating point pattern
+	* `squeeze` for choosing between list of patterns with a pattern of integers, where patterns are squeezed into the integer event duration
+	* `splice` for choosing between slices of a pattern, where the slices are squeezed into event duration
+	* Ord and Eq instances for value type @bgold-cosmos
+	* `trigger` - support for resetting envelopes on evaluation
+	* Support for rational event values
+	* Tweak how `*>` and `<*` deal with analog patterns
+	* Caribiner link bridge support
+
+## 1.0.10 - This machine also kills fascists
+* Add exports to Sound.Tidal.Scales for `getScale` and `scaleTable`
+
+## 1.0.9 - This machine kills fascists
+* sec and msec functions for converting from seconds to cycles (for stut etc) @yaxu
+* template haskell upper bounds @yaxu
+* fix for multi-laptop sync/tempo sharing @yaxu
+* fix toScale so it doesn't break on empty lists @bgold-cosmos
+* `deconstruct` function for displaying patterns stepwise @yaxu
+* `djf` control ready for new superdirt dj filter @yaxu
+* `getScale` for handrolling/adding scales to `scale` function	* Add `djf` control for upcoming superdirt dj filter @yaxu
+
+## 1.0.8 (trying to get back to doing these, 
+## see also https://tidalcycles.org/index.php/Changes_in_Tidal_1.0.x 
+## for earlier stuff)
+
+* Add 'to', 'toArg' and 'from' controls for new superdirt routing experiments - @telephon
+* Fixes for squeezeJoin (nee unwrap') - @bgold-cosmos
+* Simplify `cycleChoose`, it is now properly discrete (one event per cycle) - @yaxu
+* The return of `<>`, `infix alias for overlay` - @yaxu
+* Fix for `wedge` to allow 0 and 1 as first parameter  - @XiNNiW
+* Support for new spectral fx - @madskjeldgaard
+* Fix for _euclidInv - @dktr0
+* `chordList` for listing chords - @XiNNiW
+* new function `soak` - @XiNNiW
+* tempo fixes - @bgold-cosmos
+* miniTidal developments - @dktr0
+* potentially more efficient euclidean patternings - @dktr0
+* unit tests for euclid - @yaxu
+* fix for `sometimesBy` - @yaxu
+
 ## 0.9.10 (and earlier missing versions from this log)
 
 * arpg, a function to arpeggiate